Poprawki ścieżek katalogów - epub buduje się w katalogu build

+++ src/epub/create_epub.py
@@ -37,8 +37,6 @@ def add_song(PATH, list):
 
 def list_of_song(path_out):
     songs_list = os.listdir(path_out)
-    #    songs_list.sort()
-    #    songs_list.sort(key=lambda x: x.title)
     list_od_meta = []
     for song in songs_list:
         if song[-5:] == '.html':
@@ -48,10 +46,10 @@ def list_of_song(path_out):
     return list_od_meta
 
 
-def create_content_opf(list_of_songs_meta):
-    TMP_PATH = 'templates/content.opf'
-    OUT_PATH = '../../epub/OEBPS/content.opf'
-    tree = etree.parse(TMP_PATH)
+def create_content_opf(list_of_songs_meta, target_dir):
+    tmp_path = 'templates/content.opf'
+    out_path = os.path.join(target_dir, "epub", "OEBPS", "content.opf")
+    tree = etree.parse(tmp_path)
     root = tree.getroot()
     metadata = root.getchildren()[0]
     title = metadata.getchildren()[0]
@@ -65,35 +63,35 @@ def create_content_opf(list_of_songs_meta):
         x.attrib['media-type'] = "application/xhtml+xml"
         etree.SubElement(spine, "itemref").attrib['idref'] = 'p' + str(i + 1)
     et = etree.ElementTree(root)
-    et.write(OUT_PATH, pretty_print=True, method='xml', encoding='utf-8', xml_declaration=True)
+    et.write(out_path, pretty_print=True, method='xml', encoding='utf-8', xml_declaration=True)
 
 
-def create_toc_ncx(list_of_songs_meta):
-    TMP_PATH = 'templates/toc.ncx'
-    OUT_PATH = '../../EPUB/OEBPS/toc.ncx'
-    tree = etree.parse(TMP_PATH)
+def create_toc_ncx(list_of_songs_meta, target_dir):
+    tmp_path = 'templates/toc.ncx'
+    out_path = os.path.join(target_dir, "epub", "OEBPS", "toc.ncx")
+    tree = etree.parse(tmp_path)
     root = tree.getroot()
-    docTitle = root.getchildren()[1]
-    text1 = docTitle.getchildren()[0]
+    doctitle = root.getchildren()[1]
+    text1 = doctitle.getchildren()[0]
     text1.text += str(datetime.now().strftime("%d/%m/%Y %H:%M"))
-    navMap = root.getchildren()[2]
+    navmap = root.getchildren()[2]
     for i in range(len(list_of_songs_meta)):
-        navPoint = etree.SubElement(navMap, "navPoint")
-        navPoint.attrib['id'] = 'p' + str(i + 1)
-        navPoint.attrib['playOrder'] = str(i + 1)
-        navLabel = etree.SubElement(navPoint, "navLabel")
-        text = etree.SubElement(navLabel, "text")
+        navpoint = etree.SubElement(navmap, "navPoint")
+        navpoint.attrib['id'] = 'p' + str(i + 1)
+        navpoint.attrib['playOrder'] = str(i + 1)
+        navlabel = etree.SubElement(navpoint, "navLabel")
+        text = etree.SubElement(navlabel, "text")
         text.text = list_of_songs_meta[i].title
-        content = etree.SubElement(navPoint, "content")
+        content = etree.SubElement(navpoint, "content")
         content.attrib['src'] = list_of_songs_meta[i].plik[:-4] + '.html'
     et = etree.ElementTree(root)
-    et.write(OUT_PATH, pretty_print=True, method='xml', encoding='utf-8', xml_declaration=True)
+    et.write(out_path, pretty_print=True, method='xml', encoding='utf-8', xml_declaration=True)
 
 
-def create_toc_xhtml(list_of_songs_meta):
-    TMP_PATH = 'templates/toc.xhtml'
-    OUT_PATH = '../../EPUB/OEBPS/toc.xhtml'
-    tree = etree.parse(TMP_PATH)
+def create_toc_xhtml(list_of_songs_meta, target_dir):
+    tmp_path = 'templates/toc.xhtml'
+    out_path = os.path.join(target_dir, "epub", "OEBPS", "toc.xhtml")
+    tree = etree.parse(tmp_path)
     root = tree.getroot()
     body = root.getchildren()[1]
     nav = body.getchildren()[0]
@@ -104,12 +102,13 @@ def create_toc_xhtml(list_of_songs_meta):
         a.attrib['href'] = list_of_songs_meta[i].plik[:-4] + '.html'
         a.text = list_of_songs_meta[i].title
     et = etree.ElementTree(root)
-    et.write(OUT_PATH, pretty_print=True, method='xml', encoding='utf-8', xml_declaration=True)
+    et.write(out_path, pretty_print=True, method='xml', encoding='utf-8', xml_declaration=True)
 
 
 def create_template_epub(path):
     path_epub = os.path.join(path, "epub")
-    shutil.rmtree(path_epub)
+    if os.path.exists(path_epub):
+        shutil.rmtree(path_epub)
     path_meta = os.path.join(path_epub, "META-INF")
     path_oebps = os.path.join(path_epub, "OEBPS")
     path_css = os.path.join(path_oebps, "CSS")
@@ -127,38 +126,38 @@ def create_template_epub(path):
     shutil.copyfile(path_tmp_mimetype, os.path.join(path_epub, "mimetype"))
 
 
-def create_full_epub(src_of_songs, src, direct):
-    create_template_epub(direct)
-    path_out = os.path.join(direct, "epub", "OEBPS")
+def create_full_epub(src_of_songs, src, target_dir):
+    create_template_epub(target_dir)
+    path_out = os.path.join(target_dir, "epub", "OEBPS")
     cash.create_all_songs_html(src_of_songs, src, path_out)
-    create_content_opf(list_of_song(path_out))
-    create_toc_ncx(list_of_song(path_out))
-    create_toc_xhtml(list_of_song(path_out))
+    create_content_opf(list_of_song(path_out), target_dir)
+    create_toc_ncx(list_of_song(path_out), target_dir)
+    create_toc_xhtml(list_of_song(path_out), target_dir)
 
 
-def package_epub(direct):
-    direct_epub = os.path.join(direct, "epub")
-    with ZipFile(os.path.join(direct, "spiewnik.epub"), 'w', compression=zipfile.ZIP_DEFLATED) as myzip:
-        myzip.write(os.path.join(direct_epub, "mimetype"), arcname="mimetype", compress_type=zipfile.ZIP_STORED)
-        myzip.write(os.path.join(direct_epub, "META-INF", "container.xml"),
+def package_epub(target_dir):
+    target_dir_epub = os.path.join(target_dir, "epub")
+    with ZipFile(os.path.join(target_dir, "spiewnik.epub"), 'w', compression=zipfile.ZIP_DEFLATED) as myzip:
+        myzip.write(os.path.join(target_dir_epub, "mimetype"), arcname="mimetype", compress_type=zipfile.ZIP_STORED)
+        myzip.write(os.path.join(target_dir_epub, "META-INF", "container.xml"),
                     arcname=os.path.join("META-INF", "container.xml"))
-        file_list = os.listdir(os.path.join(direct_epub, "OEBPS"))
+        file_list = os.listdir(os.path.join(target_dir_epub, "OEBPS"))
         for file in file_list:
             if file != 'CSS':
-                myzip.write(os.path.join(direct_epub, "OEBPS", file), arcname=os.path.join("OEBPS", file))
-        myzip.write(os.path.join(direct_epub, "OEBPS", "CSS", "template.css"),
+                myzip.write(os.path.join(target_dir_epub, "OEBPS", file), arcname=os.path.join("OEBPS", file))
+        myzip.write(os.path.join(target_dir_epub, "OEBPS", "CSS", "template.css"),
                     arcname=os.path.join("OEBPS", "CSS", "template.css"))
-        myzip.write(os.path.join(direct_epub, "OEBPS", "CSS", "song.css"),
+        myzip.write(os.path.join(target_dir_epub, "OEBPS", "CSS", "song.css"),
                     arcname=os.path.join("OEBPS", "CSS", "song.css"))
 
 
 def main():
     src = os.path.join("..", "..", "songs")  # gdzie są wszystkie piosenki
-    direct = os.path.join("..", "..")  # gdzie ma utworzyć epub
+    target_dir = os.path.join("..", "..", "build")  # gdzie ma utworzyć epub
     src_of_songs = os.path.join("..", "..", "songs")
     # które piosenki chcę zawrzeć w śpiewniku(może być katalogiem z plikami xml lub listą plików)
-    create_full_epub(src_of_songs, src, direct)
-    package_epub(direct)
+    create_full_epub(src_of_songs, src, target_dir)
+    package_epub(target_dir)
 
 
 if __name__ == "__main__":
